Fix home tab view model bugs
