+++ WPF/VMagicMirrorConfig/ViewModel/ControlPanel/HomeViewModel.cs
@@ -1,7 +1,10 @@
 ﻿using Baku.VMagicMirrorConfig.View;
+using MahApps.Metro.Controls;
+using MahApps.Metro.Controls.Dialogs;
 using Microsoft.Win32;
 using System.Collections.ObjectModel;
 using System.IO;
+using System.Threading.Tasks;
 using System.Windows;
 
 namespace Baku.VMagicMirrorConfig.ViewModel
@@ -12,13 +15,13 @@ namespace Baku.VMagicMirrorConfig.ViewModel
     public class HomeViewModel : ViewModelBase
     {
         private readonly RootSettingModel _setting;
-        private readonly SettingFileIo _settingFileIo;
+        private readonly SaveFileManager _saveFileManager;
         private readonly AvatarLoader _avatarLoader;
         private readonly AppQuitSetting _appQuitSetting;
 
         public HomeViewModel() : this(
             ModelResolver.Instance.Resolve<RootSettingModel>(),
-            ModelResolver.Instance.Resolve<SettingFileIo>(),
+            ModelResolver.Instance.Resolve<SaveFileManager>(),
             ModelResolver.Instance.Resolve<AvatarLoader>(),
             ModelResolver.Instance.Resolve<ScreenshotTaker>(),
             ModelResolver.Instance.Resolve<AppQuitSetting>()
@@ -28,14 +31,14 @@ namespace Baku.VMagicMirrorConfig.ViewModel
 
         internal HomeViewModel(
             RootSettingModel rootSetting, 
-            SettingFileIo settingFileIo,
+            SaveFileManager saveFileManager,
             AvatarLoader avatarLoader,
             ScreenshotTaker screenshotTaker,
             AppQuitSetting appQuitSetting
             )
         {
             _setting = rootSetting;
-            _settingFileIo = settingFileIo;
+            _saveFileManager = saveFileManager;
             _avatarLoader = avatarLoader;
             _appQuitSetting = appQuitSetting;
             
@@ -47,12 +50,18 @@ namespace Baku.VMagicMirrorConfig.ViewModel
             AutoAdjustCommand = new ActionCommand(() => avatarLoader.RequestAutoAdjust());
             OpenSettingWindowCommand = new ActionCommand(() => SettingWindow.OpenOrActivateExistingWindow());
 
-            ResetToDefaultCommand = new ActionCommand(ResetToDefault);
-
             TakeScreenshotCommand = new ActionCommand(() => screenshotTaker.TakeScreenshot());
             OpenScreenshotFolderCommand = new ActionCommand(() => screenshotTaker.OpenScreenshotSavedFolder());
 
-            if (!IsInDesignMode)
+            ShowSaveModalCommand = new ActionCommand(ShowSaveModal);
+            ShowLoadModalCommand = new ActionCommand(ShowLoadModal);
+
+            ExportSettingToFileCommand = new ActionCommand(SaveSettingToFile);
+            ImportSettingFromFileCommand = new ActionCommand(LoadSettingFromFile);
+
+            ResetToDefaultCommand = new ActionCommand(ResetToDefault);
+
+            if (IsInDesignMode)
             {
                 return;
             }
@@ -63,6 +72,7 @@ namespace Baku.VMagicMirrorConfig.ViewModel
             if (_activateOnStartup)
             {
                 RaisePropertyChanged(nameof(ActivateOnStartup));
+                LogOutput.Instance.Write($"prop changed activate on startup");
             }
             OtherVersionRegisteredOnStartup = regSetting.CheckOtherVersionRegistered();
         }
@@ -83,6 +93,7 @@ namespace Baku.VMagicMirrorConfig.ViewModel
                 if (SetValue(ref _activateOnStartup, value))
                 {
                     new StartupRegistrySetting().SetThisVersionRegister(value);
+                    LogOutput.Instance.Write($"Register this version to startup, {value}");
                     if (value)
                     {
                         OtherVersionRegisteredOnStartup = false;
@@ -153,8 +164,128 @@ namespace Baku.VMagicMirrorConfig.ViewModel
             //書いてる通りだが、「設定のデフォルト化 == 設定ファイルを消してオートセーブが無効な状態で再起動」とすることで
             //クリーンに再起動しようとしている
             _appQuitSetting.SkipAutoSaveAndRestart = true;
-            _settingFileIo.DeleteSetting(SpecialFilePath.AutoSaveSettingFilePath);
+            _saveFileManager.SettingFileIo.DeleteSetting(SpecialFilePath.AutoSaveSettingFilePath);
             Application.Current.MainWindow?.Close();
         }
+
+        #region セーブ/ロード
+
+        public ActionCommand ShowSaveModalCommand { get; }
+        public ActionCommand ShowLoadModalCommand { get; }
+
+        private async void ShowSaveModal()
+        {
+            if (Application.Current.MainWindow is not MetroWindow window)
+            {
+                return;
+            }
+
+            var progress = await GuardSettingWindowIfNeeded();
+
+            var dialog = new SaveLoadMetroDialog();
+            var vm = SaveLoadDataViewModel.CreateForSave(_saveFileManager, async () =>
+            {
+                await window.HideMetroDialogAsync(dialog);
+                if (progress != null)
+                {
+                    await progress.CloseAsync();
+                }
+            });
+
+            dialog.DataContext = vm;
+            await window.ShowMetroDialogAsync(dialog, new MetroDialogSettings()
+            {
+                AnimateShow = true,
+                AnimateHide = false,
+                OwnerCanCloseWithDialog = true,
+            });
+            await dialog.WaitUntilUnloadedAsync();
+        }
+
+        private async void ShowLoadModal()
+        {
+            if (Application.Current.MainWindow is not MetroWindow window)
+            {
+                return;
+            }
+
+            var progress = await GuardSettingWindowIfNeeded();
+
+            var dialog = new SaveLoadMetroDialog();
+            var vm = SaveLoadDataViewModel.CreateForLoad(_setting, _saveFileManager, async () =>
+            {
+                await window.HideMetroDialogAsync(dialog);
+                if (progress != null)
+                {
+                    await progress.CloseAsync();
+                }
+            });
+
+            dialog.DataContext = vm;
+            await window.ShowMetroDialogAsync(dialog, new MetroDialogSettings()
+            {
+                AnimateShow = true,
+                AnimateHide = false,
+                OwnerCanCloseWithDialog = true,
+            });
+            await dialog.WaitUntilUnloadedAsync();
+        }
+
+        private async Task<ProgressDialogController?> GuardSettingWindowIfNeeded()
+        {
+            if (SettingWindow.CurrentWindow is not SettingWindow settingWindow)
+            {
+                return null;
+            }
+
+            var indication = MessageIndication.GuardSettingWindowDuringSaveLoad();
+            return await settingWindow.ShowProgressAsync(indication.Title, indication.Content, settings: new MetroDialogSettings()
+            {
+                DialogResultOnCancel = MessageDialogResult.Negative,
+                AnimateShow = true,
+                AnimateHide = false,
+                OwnerCanCloseWithDialog = true,
+            });
+        }
+
+        #endregion
+
+        #region エクスポート/インポート
+
+        public ActionCommand ExportSettingToFileCommand { get; }
+        public ActionCommand ImportSettingFromFileCommand { get; }
+
+        private void SaveSettingToFile()
+        {
+            var dialog = new SaveFileDialog()
+            {
+                Title = "Save VMagicMirror Setting",
+                Filter = "VMagicMirror Setting File(*.vmm)|*.vmm",
+                DefaultExt = ".vmm",
+                AddExtension = true,
+            };
+            if (dialog.ShowDialog() == true)
+            {
+                _saveFileManager.SettingFileIo.SaveSetting(dialog.FileName, SettingFileReadWriteModes.Exported);
+                SnackbarWrapper.Enqueue(LocalizedString.GetString("SettingFile_SaveCompleted_ExportedFile"));
+            }
+        }
+
+        private void LoadSettingFromFile()
+        {
+            var dialog = new OpenFileDialog()
+            {
+                Title = "Load VMagicMirror Setting",
+                Filter = "VMagicMirror Setting File (*.vmm)|*.vmm",
+                Multiselect = false,
+            };
+            if (dialog.ShowDialog() == true)
+            {
+                _saveFileManager.SettingFileIo.LoadSetting(dialog.FileName, SettingFileReadWriteModes.Exported);
+                SnackbarWrapper.Enqueue(LocalizedString.GetString("SettingFile_LoadCompleted_ExportedFile"));
+            }
+        }
+
+        #endregion
     }
 }
