+++ WPF/VMagicMirrorConfig/ViewModel/SettingWindowTab/SettingIoViewModel.cs
@@ -1,41 +1,23 @@
-﻿using Baku.VMagicMirrorConfig.View;
-using MahApps.Metro.Controls;
-using MahApps.Metro.Controls.Dialogs;
-using Microsoft.Win32;
-using System.ComponentModel;
-using System.Threading.Tasks;
-using System.Windows;
+﻿using System.ComponentModel;
 
 namespace Baku.VMagicMirrorConfig.ViewModel
 {
     public class SettingIoViewModel : SettingViewModelBase
     {
         public SettingIoViewModel() : this(
-            ModelResolver.Instance.Resolve<RootSettingModel>(),
-            ModelResolver.Instance.Resolve<AutomationSettingModel>(),
-            ModelResolver.Instance.Resolve<SaveFileManager>()
+            ModelResolver.Instance.Resolve<AutomationSettingModel>()
             )
         {
         }
 
-        internal SettingIoViewModel(
-            RootSettingModel rootModel, AutomationSettingModel model, SaveFileManager saveFileManager
-            )
+        internal SettingIoViewModel(AutomationSettingModel model)
         {
-            _rootModel = rootModel;
             _model = model;
-            _saveFileManager = saveFileManager;
 
             OpenInstructionUrlCommand = new ActionCommand(OpenInstructionUrl);
             RequestEnableAutomationCommand = new ActionCommand(OnEnableAutomationRequested);
             RequestDisableAutomationCommand = new ActionCommand(OnDisableAutomationRequested);
             ApplyPortNumberCommand = new ActionCommand(ApplyPortNumber);
-
-            ShowSaveModalCommand = new ActionCommand(ShowSaveModal);
-            ShowLoadModalCommand = new ActionCommand(ShowLoadModal);
-
-            ExportSettingToFileCommand = new ActionCommand(SaveSettingToFile);
-            ImportSettingFromFileCommand = new ActionCommand(LoadSettingFromFile);
 
             if (!IsInDesignMode)
             {
@@ -53,140 +35,7 @@ namespace Baku.VMagicMirrorConfig.ViewModel
             _model.AutomationPortNumber.AddWeakEventHandler(OnAutomationPortNumberChanged);
         }
 
-        //NOTE: rootが必要なのは「ロード時にキャラ情報/非キャラ情報をどう扱うか」という値がRootに入っているため。
-        //コレ以外の目的で使うのは濫用に当たるので注意
-        private readonly RootSettingModel _rootModel;
         private readonly AutomationSettingModel _model;
-        private readonly SaveFileManager _saveFileManager;
-
-
-        private void OnAutomationPortNumberChanged(object? sender, PropertyChangedEventArgs e)
-        {
-            AutomationPortNumberText.Value = _model.AutomationPortNumber.Value.ToString();
-        }
-
-        #region セーブ/ロード
-
-        public ActionCommand ShowSaveModalCommand { get; }
-        public ActionCommand ShowLoadModalCommand { get; }
-
-        private async void ShowSaveModal()
-        {
-            if (Application.Current.MainWindow is not MetroWindow window)
-            {
-                return;
-            }
-
-            var progress = await GuardSettingWindowIfNeeded();
-
-            var dialog = new SaveLoadMetroDialog();
-            var vm = SaveLoadDataViewModel.CreateForSave(_saveFileManager, async () =>
-            {
-                await window.HideMetroDialogAsync(dialog);
-                if (progress != null)
-                {
-                    await progress.CloseAsync();
-                }
-            });
-
-            dialog.DataContext = vm;
-            await window.ShowMetroDialogAsync(dialog, new MetroDialogSettings()
-            {
-                AnimateShow = true,
-                AnimateHide = false,
-                OwnerCanCloseWithDialog = true,
-            });
-            await dialog.WaitUntilUnloadedAsync();
-        }
-
-        private async void ShowLoadModal()
-        {
-            if (Application.Current.MainWindow is not MetroWindow window)
-            {
-                return;
-            }
-
-            var progress = await GuardSettingWindowIfNeeded();
-
-            var dialog = new SaveLoadMetroDialog();
-            var vm = SaveLoadDataViewModel.CreateForLoad(_rootModel, _saveFileManager, async () =>
-            {
-                await window.HideMetroDialogAsync(dialog);
-                if (progress != null)
-                {
-                    await progress.CloseAsync();
-                }
-            });
-
-            dialog.DataContext = vm;
-            await window.ShowMetroDialogAsync(dialog, new MetroDialogSettings()
-            {
-                AnimateShow = true,
-                AnimateHide = false,
-                OwnerCanCloseWithDialog = true,
-            });
-            await dialog.WaitUntilUnloadedAsync();
-        }
-
-        private async Task<ProgressDialogController?> GuardSettingWindowIfNeeded()
-        {
-            if (SettingWindow.CurrentWindow is not SettingWindow settingWindow)
-            {
-                return null;
-            }
-
-            var indication = MessageIndication.GuardSettingWindowDuringSaveLoad();
-            return await settingWindow.ShowProgressAsync(indication.Title, indication.Content, settings: new MetroDialogSettings()
-            {
-                DialogResultOnCancel = MessageDialogResult.Negative,
-                AnimateShow = true,
-                AnimateHide = false,
-                OwnerCanCloseWithDialog = true,
-            });            
-        }
-
-        #endregion
-
-
-        #region エクスポート/インポート
-
-        public ActionCommand ExportSettingToFileCommand { get; }
-        public ActionCommand ImportSettingFromFileCommand { get; }
-
-        private void SaveSettingToFile()
-        {
-            var dialog = new SaveFileDialog()
-            {
-                Title = "Save VMagicMirror Setting",
-                Filter = "VMagicMirror Setting File(*.vmm)|*.vmm",
-                DefaultExt = ".vmm",
-                AddExtension = true,
-            };
-            if (dialog.ShowDialog() == true)
-            {
-                _saveFileManager.SettingFileIo.SaveSetting(dialog.FileName, SettingFileReadWriteModes.Exported);
-                SnackbarWrapper.Enqueue(LocalizedString.GetString("SettingFile_SaveCompleted_ExportedFile"));
-            }
-        }
-
-        private void LoadSettingFromFile()
-        {
-            var dialog = new OpenFileDialog()
-            {
-                Title = "Load VMagicMirror Setting",
-                Filter = "VMagicMirror Setting File (*.vmm)|*.vmm",
-                Multiselect = false,
-            };
-            if (dialog.ShowDialog() == true)
-            {
-                _saveFileManager.SettingFileIo.LoadSetting(dialog.FileName, SettingFileReadWriteModes.Exported);
-                SnackbarWrapper.Enqueue(LocalizedString.GetString("SettingFile_LoadCompleted_ExportedFile"));
-            }
-        }
-
-        #endregion
-
-        #region オートメーションっぽい所
 
         public RProperty<bool> IsAutomationEnabled => _model.IsAutomationEnabled;
 
@@ -238,7 +87,9 @@ namespace Baku.VMagicMirrorConfig.ViewModel
         private void OpenInstructionUrl()
             => UrlNavigate.Open(LocalizedString.GetString("URL_docs_setting_files"));
 
-        #endregion
-
+        private void OnAutomationPortNumberChanged(object? sender, PropertyChangedEventArgs e)
+        {
+            AutomationPortNumberText.Value = _model.AutomationPortNumber.Value.ToString();
+        }
     }
 }
